Fix circular import of modules needed for type-checking
