+++ src/gt4py/next/program_processors/runners/dace_fieldview/gtir_dace_backend.py
@@ -12,6 +12,8 @@
 #
 # SPDX-License-Identifier: GPL-3.0-or-later
 
+from __future__ import annotations
+
 import dace
 
 from gt4py.next import common as gtx_common
