+++ src/gt4py/next/program_processors/runners/dace_fieldview/gtir_python_codegen.py
@@ -12,6 +12,8 @@
 #
 # SPDX-License-Identifier: GPL-3.0-or-later
 
+from __future__ import annotations
+
 from typing import Any
 
 import numpy as np
