+++ src/gt4py/next/program_processors/runners/dace_fieldview/gtir_to_tasklet.py
@@ -13,6 +13,8 @@
 # SPDX-License-Identifier: GPL-3.0-or-later
 
 
+from __future__ import annotations
+
 import itertools
 from dataclasses import dataclass
 from typing import Optional, TypeAlias
